Added Lab 7 Asm 5 (#5)
Lab 7: Assignment 5

----------------------------------------

    Assignment Deadline: 23 March 2023, 11:59PM
    Submission:
    1. cpp file with student ID and name
    2. a screenshot of result and code

----------------------------------------

Problem: Function and Prototype
Problem description

Write a program that accepts a sequence of integers. All the input values should be stored in an array.  In this task, the maximum length of the sequence is 1000. Declare the array with an appropriate size accordingly. 

Develop 4 functions by yourself to process an array and output the largest and smallest numbers, and the total and average of all numbers. Specify 4 corresponding function prototypes

For the output format of average, please specify 3 digits after the decimal point.

Similar to the previous assignment, users will use -999 to indicate the end of input. Sample inputs and outputs are as follows.

Sample Input and Output 1:
Enter a sequence of integer (-999 to finish): 1 2 3 4 5 6 7 8 9 -999
Largest Number is 9
Smallest Number is 1
Total is 45
Average is 5.000